from os/user to github.com/mitchellh/go-homedir
Why not just use os/user? The built-in os/user package requires cgo on some operatinal systems. This means that any Go code that uses that package cannot cross compile. But 99% of the time the use for os/user is just to retrieve the home directory, which we can do for the current user without cgo. This library does that, enabling cross-compilation